dpif-netdev: Allow PMD auto load balance with cross-numa.
Previously, auto load balance did not trigger a reassignment when
there was cross-numa polling as an rxq could be polled from a
different numa after reassign and it could impact estimates.

In the case where there is only one numa with pmds available, the
same numa will always poll before and after reassignment, so estimates
are safe. Allow PMD auto load balance to trigger a reassignment in
this case.

Signed-off-by: Kevin Traynor <ktraynor@redhat.com>