dpif-netdev: Restructure rxq schedule logging.
Previously logging about rxq scheduling was done in a code branch with
the selection of the PMD thread core after checking that a numa was
selected.

By splitting out the logging from the PMD thread core selection, it can
simplify the code complexity and make it more extendable for future
additions.

Also, minor updates to variables to improve readability and fix an
incorrect log indent.

There is no user visibile change in behaviour or logs.

Signed-off-by: Kevin Traynor <ktraynor@redhat.com>