dpdk: Use DPDK 20.11.6 release.
Update OVS CLI and relevant documentation to use DPDK 20.11.6.

A bug was introduced in DPDK 20.11.5 by the commit
33f2e3756186 ("vhost: fix unsafe vring addresses modifications").
This bug can cause a deadlock when vIOMMU is enabled and
NUMA reallocation of the virtqueues happen.
A fix [1] has been posted and pushed to the DPDK 20.11 branch.
If a user wishes to avoid the issue then it is recommended to use
DPDK 20.11.4 until the release of DPDK 20.11.7.
It should be noted that DPDK 20.11.4 does not benefit from
the numerous bug fixes addressed since its release.
If a user wishes to benefit from these fixes it is recommended
to use DPDK 20.11.6.

[1] https://patches.dpdk.org/project/dpdk/patch/20220725203206.427083-2-david.marchand@redhat.com/

Signed-off-by: Michael Phelan <michael.phelan@intel.com>