dpif-netdev: Add PMD sleep start delay.
It may be desirable to wait some time of time before sleeping
to avoid wakeup latencies.

Add the ability for the user to delay the PMD from starting to
sleep until a configured time has expired.

It will only delay the initial sleep while sleep conditions
are met and will be reset when the sleep conditions are
no longer present.

To set the time in microseconds, use:
ovs-vsctl set Open_vSwitch . other_config:pmd-delaysleep=1000

By default it is set to 0, so there is no change in default behaviour.

Signed-off-by: Kevin Traynor <ktraynor@redhat.com>