netdev-dpdk: Check if device is pending reset when added.
When a device reset interrupt event is detected for a DPDK device
added to OVS, the device reset is performed.

If a device reset interrupt event is detected for a device before
it is added to OVS, device reset is not called.

If that device is later attempted to be added to OVS, it may fail
while being configured as it is still pending a reset. This was
observed with the DPDK iavf driver, by forcing a reset event from
the ice driver and then adding the iavf device to OVS.

Reset was not previously called during initial device configuration.

Add a check if there was any previous device reset interrupt events
when a device is added to OVS. If there was, perform the reset
before continuing with configuration.

Signed-off-by: Kevin Traynor <ktraynor@redhat.com>