dpdk: Check other_config:dpdk-extra for '--lcores'.
Currently dpdk lcore args for DPDK EAL init can be generated or it
can be written directly by the user through dpdk-extra.

If dpdk-extra does not contain '-l' or '-c', a '-l' argument with
a core list for DPDK EAL init will be generated.

The '--lcores' argument should also be check, as if it is used in
dpdk-extra, a '-l' is still generated and that causes DPDK EAL init
to fail.

Add check for '--lcores' in dpdk-extra config.

Fixes: 0a0f39df1d5a ("netdev-dpdk: Add support for DPDK 16.07")
Signed-off-by: Kevin Traynor <ktraynor@redhat.com>