appveyor: More robust OpenSSL installation wrt path changes.
AppVeyor had an issue where C:\OpenSSL-Win64 folder didn't exist for
some reason and the installation of OpenSSL was placing it into
C:\Program Files\OpenSSL-Win64 instead as well.  After that we couldn't
find the libraries and the build failed:
  https://help.appveyor.com/discussions/problems/38517-cannot-find-path-copenssl-win64-because-it-does-not-exist

The issue was resolved since, but it's better if we check the paths
and install into specific location to be more resilient to this kind
of environment issues, i.e. make less assumptions.

While at it, fixing the slash type for the path.  Remove-Item somehow
accepts the "wrong" one, but a backslash is more native in paths on
Windows.  We use forward slash while in msys2 shell, but should not use
it in PowerShell environment.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>